+++ tutorial/1-runtime/1-minimal.cc
@@ -1,4 +1,5 @@
 #include <flecsi/execution.hh>
+#include <flecsi/run/control.hh>
 
 /*
   The top-level action can be any C/C++ function that takes no arguments and
@@ -17,34 +18,20 @@ top_level_action() {
 } // top_level_action
 
 /*
-  The main function must invoke initialize, start, and finalize on the FleCSI
-  runtime. Otherwise, the implementation of main is left to the user.
+  The main function must create a FleCSI \c runtime object.  Otherwise, the
+  implementation of main is left to the user.
  */
 
 int
 main(int argc, char ** argv) {
-
-  auto status = flecsi::initialize(argc, argv);
-
+  flecsi::run::arguments args(argc, argv);
+  const flecsi::run::dependencies_guard dg(args.dep);
+  const flecsi::runtime run(args.cfg);
   /*
-    The status returned by FleCSI's initialize method should be inspected to
-    see if the end-user specified --help on the command line. FleCSI has
-    built-in command-line support using Boost Program Options. This is
-    documented in the next example.
+    flecsi::run::call means to call the single function given as an argument to
+    run.main.  (Multiple functions will be covered in later examples.)
+    It will be skipped if, say, --help was passed as an argument; FleCSI's
+    command-line support is documented in the next example.
    */
-
-  if(status != flecsi::run::status::success) {
-    return status == flecsi::run::status::help ? 0 : status;
-  } // if
-
-  /*
-    The top-level action is passed to the start function to tell FleCSI what
-    it should execute.
-   */
-
-  status = flecsi::start(top_level_action);
-
-  flecsi::finalize();
-
-  return status;
+  return run.main<flecsi::run::call>(args.act, top_level_action);
 } // main