+++ tutorial/1-runtime/4-caliper.cc
@@ -2,6 +2,7 @@
 #include <thread>
 
 #include <flecsi/execution.hh>
+#include <flecsi/run/control.hh>
 #include <flecsi/util/annotation.hh>
 
 using namespace flecsi;
@@ -49,17 +50,10 @@ int
 main(int argc, char * argv[]) {
   annotation::rguard<main_region> main_guard;
 
-  auto status = flecsi::initialize(argc, argv);
-
-  if(status != flecsi::run::status::success) {
-    return status == flecsi::run::status::help ? 0 : status;
-  }
-
-  status = (annotation::guard<annotation::execution, annotation::detail::low>(
-              "top-level-task"),
-    flecsi::start(top_level_action));
-
-  flecsi::finalize();
-
-  return status;
+  run::arguments args(argc, argv);
+  const run::dependencies_guard dg(args.dep);
+  const runtime run(args.cfg);
+  return (annotation::guard<annotation::execution, annotation::detail::low>(
+            "top-level-task"),
+    run.main<run::call>(args.act, top_level_action));
 } // main