+++ doc/sphinx/src/news.rst
@@ -20,6 +20,12 @@ Changes in v2.3.0
 
 Deprecated
 ^^^^^^^^^^
+* Runtime
+
+  * ``initialize``, ``start``, ``finalize``, ``run::status``, and ``control::check_status`` |mdash| use ``runtime``
+  * ``program`` |mdash| use ``run::arguments``
+  * ``control::execute`` |mdash| use ``control::invoke`` or ``runtime::main``
+
 * Utilities
 
   * Passing binary functors to ``mpi::one_to_allv``, ``mpi::one_to_alli``, and ``mpi::all_to_allv`` |mdash| remove second parameter or use ranges
@@ -28,6 +34,8 @@ New features
 ^^^^^^^^^^^^
 * Runtime
 
+  * ``run::arguments`` and its nested types represent a parsed command line.
+  * ``run::dependencies_guard`` allows for application control over initialization of FleCSI dependencies.
   * ``control::invoke`` executes a control model with a policy object constructed from arguments provided.
   * ``run::call`` is a trivial predefined control model.